better layout for chat and userlist
