+++ client/src/components/Chat/chat.css
@@ -1,29 +1,46 @@
-body {
-  margin: 0;
-  padding-bottom: 3rem;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica,
-    Arial, sans-serif;
+#chat-box {
+  background-color: #FDE1B1;
+  border: 3px solid #ACA7DF;
+  border-radius: 1rem;
+  height: 20rem;
+  padding: 0.25rem;
+}
+
+.chat-wrapper {
+  /* display: flex;
+  flex-direction: column-reverse; */
+  height: 17rem;
+  overflow-y: scroll;
+  overscroll-behavior-y: contain;
+  scroll-snap-type: y proximity;
+  
+}
+
+.chat-msg {
+  border-radius: 1rem;
+}
+
+
+#chat-input {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
 }
 
 #chat-form {
   background: rgba(0, 0, 0, 0.15);
-  padding: 0.25rem;
-  position: absolute;
-  bottom: 64px;
-  left: 0;
-  right: 0;
-  display: flex;
   height: 3rem;
-  box-sizing: border-box;
   backdrop-filter: blur(10px);
+  border-radius: 1rem;
 }
 
 #input {
   border: none;
-  padding: 0 1rem;
-  flex-grow: 1;
+  padding: 0 0.75rem;
   border-radius: 2rem;
   margin: 0.25rem;
+  margin-left: 1.5rem;
 }
 
 #input:focus {
@@ -33,7 +50,7 @@ body {
   background-color: #87D8EE;
   border: none;
   padding: 0 1rem;
-  margin: 0.25rem;
+  margin: 0.25rem 0rem 0.25rem;
   border-radius: 2rem;
   outline: none;
   color: white;
@@ -44,7 +61,6 @@ body {
   padding: 0;
   font-size: 1rem;
   line-height: 1.25rem;
-  height: 200px;
 }
 
 #messages > li {
@@ -55,12 +71,9 @@ body {
   background: rgb(228, 244, 250);
 }
 
-#chat-box {
-  overflow: scroll;
+#messages > li:last-child {
+  scroll-snap-align: end;
 }
 
-.chat-msg {
-  height: 200px;
-}
